requirements.txt

streamlit
streamlit-drawable-canvas
Pillow
torchvision
scikit-learn
